+++ compiler/src/main/java/net/zerobuilder/Builders.java
@@ -5,6 +5,7 @@ import java.lang.annotation.Target;
 
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.SOURCE;
+import static net.zerobuilder.AccessLevel.PUBLIC;
 
 /**
  * <p>
@@ -25,5 +26,6 @@ import static java.lang.annotation.RetentionPolicy.SOURCE;
 public @interface Builders {
 
   boolean recycle() default false;
+  AccessLevel access() default PUBLIC;
 
 }