+++ compiler/src/main/java/net/zerobuilder/Goal.java
@@ -7,7 +7,7 @@ import static java.lang.annotation.ElementType.CONSTRUCTOR;
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.SOURCE;
-import static net.zerobuilder.AccessLevel.PUBLIC;
+import static net.zerobuilder.AccessLevel.UNSPECIFIED;
 
 /**
  * <p>
@@ -75,6 +75,6 @@ public @interface Goal {
    */
   boolean nonNull() default false;
 
-  AccessLevel builderAccess() default PUBLIC;
-  AccessLevel toBuilderAccess() default PUBLIC;
+  AccessLevel builderAccess() default UNSPECIFIED;
+  AccessLevel toBuilderAccess() default UNSPECIFIED;
 }