+++ examples/basic/src/test/java/net/zerobuilder/examples/beans/more/ExperimentsTest.java
@@ -1,5 +1,6 @@
 package net.zerobuilder.examples.beans.more;
 
+import net.zerobuilder.examples.beans.more.Experiments.Access;
 import net.zerobuilder.examples.beans.more.Experiments.AeroExperiment;
 import net.zerobuilder.examples.beans.more.Experiments.BioExperiment;
 import net.zerobuilder.examples.beans.more.Experiments.Ignorify;
@@ -7,11 +8,15 @@ import net.zerobuilder.examples.beans.more.Experiments.IterableExperiment;
 import net.zerobuilder.examples.beans.more.Experiments.RawExperiment;
 import org.junit.Test;
 
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.List;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
+import static net.zerobuilder.examples.beans.more.Experiments_AccessBuilders.accessBuilder;
+import static net.zerobuilder.examples.beans.more.Experiments_AccessBuilders.accessToBuilder;
 import static net.zerobuilder.examples.beans.more.Experiments_AeroExperimentBuilders.aeroExperimentBuilder;
 import static net.zerobuilder.examples.beans.more.Experiments_AeroExperimentBuilders.aeroExperimentToBuilder;
 import static net.zerobuilder.examples.beans.more.Experiments_BioExperimentBuilders.bioExperimentBuilder;
@@ -23,7 +28,9 @@ import static net.zerobuilder.examples.beans.more.Experiments_IterableExperiment
 import static net.zerobuilder.examples.beans.more.Experiments_RawExperimentBuilders.rawExperimentBuilder;
 import static net.zerobuilder.examples.beans.more.Experiments_RawExperimentBuilders.rawExperimentToBuilder;
 import static org.hamcrest.core.Is.is;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 public class ExperimentsTest {
 
@@ -98,5 +105,20 @@ public class ExperimentsTest {
     assertThat(something.getThings().get(0),
         is((Iterable<String>) singletonList("a")));
     assertThat(nothing2.getThings().size(), is(0));
+  }
+
+  @Test
+  public void accessTest() throws NoSuchMethodException {
+    Access foo = accessBuilder()
+        .foo("foo");
+    Access bar = accessToBuilder(foo)
+        .foo("bar")
+        .build();
+    Method builderMethod = Experiments_AccessBuilders.class.getDeclaredMethod("accessBuilder");
+    Method toBuilderMethod = Experiments_AccessBuilders.class.getDeclaredMethod("accessToBuilder", Access.class);
+    assertFalse(Modifier.isPublic(builderMethod.getModifiers()));
+    assertTrue(Modifier.isPublic(toBuilderMethod.getModifiers()));
+    assertThat(foo.getFoo(), is("foo"));
+    assertThat(bar.getFoo(), is("bar"));
   }
 }